requirements.txt
asgiref==3.8.1
Django==5.1.5
django-environ==0.12.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
phonenumbers==8.13.54
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
